+++ ttnn/cpp/ttnn/operations/data_movement/CMakeLists.txt
@@ -150,7 +150,8 @@ target_sources(
         permute/device/permute_rm_program_factory.cpp
         permute/device/permute_tiled_program_factory.cpp
         permute/permute.cpp
-        repeat/device/repeat_program_factory.cpp
+        repeat/device/repeat_program_factory_second_dim.cpp
+        repeat/device/repeat_program_factory_last_dim.cpp
         repeat/device/repeat_device_operation.cpp
         repeat/repeat.cpp
         repeat_interleave/repeat_interleave.cpp
