+++ ttnn/cpp/ttnn/operations/data_movement/repeat/device/repeat_device_operation.hpp
@@ -7,7 +7,8 @@
 #include "ttnn/run_operation.hpp"
 #include "ttnn/decorators.hpp"
 #include "ttnn/operations/data_movement/repeat/device/repeat_device_operation_types.hpp"
-#include "ttnn/operations/data_movement/repeat/device/repeat_program_factory.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_program_factory_second_dim.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_program_factory_last_dim.hpp"
 
 namespace ttnn::operations::data_movement::repeat {
