+++ ttnn/cpp/ttnn/operations/data_movement/repeat/device/repeat_program_factory_common.hpp
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <tt-metalium/core_coord.hpp>
+
+namespace ttnn::operations::data_movement::repeat::program {
+
+struct RepeatSharedVariables {
+    tt::tt_metal::KernelHandle reader_kernel_id{};
+    tt::tt_metal::CoreRange total_cores{tt::tt_metal::CoreCoord{0, 0}};
+};
+
+}  // namespace ttnn::operations::data_movement::repeat::program
